+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ New features
 ============
 
 * Add function ``modeling.create_basic_pk_model``
+* Add function ``modeling.add_metabolite``
 
 Changes
 =======